Menu refactor and year2 menu
