+++ src/year2/assets/menu/Menu-tmp.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:167441bdb8a55aa89ad2d37ee0f83555b2165550c2b911411dcdb130a775de6a
+size 7021
