+++ src/year2/components/SiteHeader.module.scss
@@ -15,16 +15,12 @@
     }
 
     @media (max-height: 450px) {
-      grid-template-columns: 1fr 1fr 1fr 1fr 1fr 1fr;
+      grid-template-columns: 1fr 1fr 1fr 1fr;
 
       & div:first-child,
       & div:nth-child(2),
       & div:nth-child(3) {
         grid-column: auto;
-      }
-
-      & div:nth-child(6) {
-        grid-column: 1;
       }
     }
   }
